refactor: reorganize file structure to keep code consistent gf-23
