+++ apps/frontend/src/libs/components/icon/libs/maps/icon-name-to-svg.map.ts
@@ -7,7 +7,7 @@ import RightArrow from "~/assets/images/icons/right-arrow.svg?react";
 import RightDoubleArrow from "~/assets/images/icons/right-double-arrow.svg?react";
 import StrikedEye from "~/assets/images/icons/striked-eye.svg?react";
 
-import { type IconName } from "../../types/types.js";
+import { type IconName } from "../types/types.js";
 
 const iconNameToSvg: Record<IconName, FC<React.SVGProps<SVGSVGElement>>> = {
 	eye: Eye,
