+++ apps/frontend/src/libs/components/icon/libs/types/icon-name.type.ts
